ArticleFormにaddArticleメソッドを移行
